segmentation_app/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-limiter==2.4.0
flasgger==0.9.7.1
opencv-python==4.5.3.56
numpy==1.21.2
Pillow==8.3.1
inference==0.9.0